+++ tests/unit/test_generate_umm_var.py
@@ -20,6 +20,16 @@ class TestGenerateUmmVar(TestCase):
         cls.launchpad_token_header = 'launchpad-token'
         cls.netcdf4_basename = 'f16_ssmis_20210426v7.nc'
         cls.netcdf4_url = f'https://example.com/{cls.netcdf4_basename}'
+        cls.query_granule_return = [
+            {
+                'links': [
+                    {
+                        'href': cls.netcdf4_url,
+                        'rel': 'http://esipfed.org/ns/fedsearch/1.1/data#',
+                    }
+                ]
+            }
+        ]
         cls.rssmif16d_variables = [
             'atmosphere_cloud_liquid_water_content',
             'atmosphere_water_vapor_content',
@@ -55,16 +65,7 @@ class TestGenerateUmmVar(TestCase):
         in detail in test_cmr_search.py.
 
         """
-        mock_granule_query.return_value.get.return_value = [
-            {
-                'links': [
-                    {
-                        'href': self.netcdf4_url,
-                        'rel': 'http://esipfed.org/ns/fedsearch/1.1/data#',
-                    }
-                ]
-            }
-        ]
+        mock_granule_query.return_value.get.return_value = self.query_granule_return
 
         # Add side effect that will copy test file to the temporary directory,
         # simulating a download.
@@ -115,16 +116,7 @@ class TestGenerateUmmVar(TestCase):
             'V0000000007-PROV',
             'V0000000008-PROV',
         ]
-        mock_granule_query.return_value.get.return_value = [
-            {
-                'links': [
-                    {
-                        'href': self.netcdf4_url,
-                        'rel': 'http://esipfed.org/ns/fedsearch/1.1/data#',
-                    }
-                ]
-            }
-        ]
+        mock_granule_query.return_value.get.return_value = self.query_granule_return
 
         # Add side effect that will copy test file to the temporary directory,
         # simulating a download.
@@ -214,16 +206,7 @@ class TestGenerateUmmVar(TestCase):
         # ['V0000000001-PROV', 'V0000000002-PROV', ..., 'Invalid JSON']
         concept_ids_and_error = concept_ids + [error_message]
 
-        mock_granule_query.return_value.get.return_value = [
-            {
-                'links': [
-                    {
-                        'href': self.netcdf4_url,
-                        'rel': 'http://esipfed.org/ns/fedsearch/1.1/data#',
-                    }
-                ]
-            }
-        ]
+        mock_granule_query.return_value.get.return_value = self.query_granule_return
 
         # Add side effect that will copy test file to the temporary directory,
         # simulating a download.
@@ -294,3 +277,58 @@ class TestGenerateUmmVar(TestCase):
 
         with self.subTest('Random string returns False'):
             self.assertFalse(is_variable_concept_id('Random string'))
+
+    @patch('varinfo.generate_umm_var.VarInfoFromNetCDF4')
+    @patch('varinfo.cmr_search.GranuleQuery')
+    @patch('varinfo.generate_umm_var.download_granule')
+    def test_generate_collection_umm_var_config_file(
+        self, mock_download_granule, mock_granule_query, mock_varinfo_from_netcdf4
+    ):
+        """This test just verifies the config file that is passed in to the
+        generate_collection_umm_var method is picked up by VarInfoFromNetCDF4 that
+        would use it. The granule query and download methods are mocked to make it
+        simpler test.
+        """
+        mock_granule_query.return_value.get.return_value = self.query_granule_return
+
+        # Add side effect that will copy test file to the temporary directory,
+        # simulating a download.
+        mock_download_granule.side_effect = self.download_granule_side_effect
+
+        # Run the test:
+        generate_collection_umm_var(
+            self.collection_concept_id,
+            self.bearer_token_header,
+            config_file='tests/unit/data/test_config.json',
+        )
+
+        # Ensure the the config file provided is passed to the VarInfo class
+        mock_varinfo_from_netcdf4.assert_called_once_with(
+            ANY,
+            config_file='tests/unit/data/test_config.json',
+        )
+
+    @patch('varinfo.generate_umm_var.VarInfoFromNetCDF4')
+    @patch('varinfo.cmr_search.GranuleQuery')
+    @patch('varinfo.generate_umm_var.download_granule')
+    def test_generate_collection_umm_var_with_no_config_file(
+        self, mock_download_granule, mock_granule_query, mock_varinfo_from_netcdf4
+    ):
+        """This test just verifies if the config file is 'None' in the
+        generate_collection_umm_var method, the VarInfoFromNetCDF4 would still succeed
+        and continue with a 'None' value for that parameter. The granule query and
+        download methods are mocked to make it simpler test.
+        """
+        mock_granule_query.return_value.get.return_value = self.query_granule_return
+
+        # Add side effect that will copy test file to the temporary directory,
+        # simulating a download.
+        mock_download_granule.side_effect = self.download_granule_side_effect
+
+        # Run the test:
+        generate_collection_umm_var(
+            self.collection_concept_id, self.bearer_token_header
+        )
+
+        # Ensure that the VarInfoFromNetCDF4 invocation includes NONE for the config file
+        mock_varinfo_from_netcdf4.assert_called_with(ANY, config_file=None)